Extract ThemeManager and rename it to ThemeController
