+++ src/settings/EditCustomColorsDialog.js
@@ -5,7 +5,7 @@ import {assertMainOrNode} from "../api/common/Env"
 import {Dialog} from "../gui/base/Dialog"
 import {ButtonType} from "../gui/base/ButtonN"
 import type {Theme} from "../gui/theme"
-import {themeManager} from "../gui/theme"
+import {themeController} from "../gui/theme"
 import type {DialogHeaderBarAttrs} from "../gui/base/DialogHeaderBar"
 import {Keys} from "../api/common/TutanotaConstants"
 import type {TextFieldAttrs} from "../gui/base/TextFieldN"
@@ -18,7 +18,7 @@ assertMainOrNode()
 let COLOR_FORMAT = new RegExp("^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$")
 
 export function show(themeToEdit: Theme, onThemeChanged: (Theme) => mixed) {
-	const colorFieldsAttrs = Object.keys(themeManager.getDefaultTheme())
+	const colorFieldsAttrs = Object.keys(themeController.getDefaultTheme())
 	                               .filter(name => name !== "logo" && name !== "themeId")
 	                               .sort((a, b) => a.localeCompare(b))
 	                               .map(colorName => {
@@ -115,13 +115,13 @@ function _getDefaultColorLine(field: TextFieldAttrs): Child {
 	if (!field.value().trim() || colorValue) {
 		let colorName = (field: any).label()
 		return m(".small.flex-space-between", [
-			m("", lang.get("defaultColor_label", {"{1}": themeManager.getDefaultTheme()[colorName]})),
+			m("", lang.get("defaultColor_label", {"{1}": themeController.getDefaultTheme()[colorName]})),
 			m("", {
 				style: {
 					width: '100px',
 					height: '10px',
 					"margin-top": "2px",
-					"background-color": themeManager.getDefaultTheme()[colorName]
+					"background-color": themeController.getDefaultTheme()[colorName]
 				}
 			})
 		])
