+++ test/client/Suite.js
@@ -8,7 +8,7 @@ import "./common/PasswordUtilsTest"
 import "./gui/animation/AnimationsTest"
 import "./gui/base/ButtonTest"
 import "./gui/base/ListTest"
-import "./gui/ThemeManagerTest"
+import "./gui/ThemeControllerTest"
 import "./common/EntropyCollectorTest"
 import "./common/HtmlSanitizerTest"
 import "./mail/InboxRuleHandlerTest"
